test(gui): the tier guards count fonts, not needles
`guard-prover` on `afe7573b` falsified that commit's own claim.
Three holes, each proven GREEN with the defect live.

**The anti-drift pair did not prevent drift.** Both tier clauses
were positive `contains`, and nothing said the shared tier was
the LAST font in the chain. Keeping
`.font(SettingsDrawerHeader.tier…)` and appending
`.font(.title3…)` to the header and `.font(.caption)` to the
summary drew 15 pt against 10 pt — maximal divergence — under a
guard whose message said the two could no longer drift apart.
SwiftUI draws the last font in a chain, so the COUNT is the
invariant and a needle is not: both runs must carry exactly one.

**A negative clause was reading a positive-only run.**
`!labelRun.contains(".font(.callout")` took its run from
`run(in:from:to:)`, which stops at the first `}` — its own
docstring says positive clauses only. One `.overlay { }` ends
the run before the literal it watches. The label's run is
brace-balanced now, and the count clause replaces the literal
one outright.

**The scan's floor did not prove a call site was scanned.** Of
the three `accessory:` closures in the tree, two are the
wrapper's own `accessory: { EmptyView() }` forwardings in
`Components/Common` — not call sites, and unable to hold a
`Text(`. Deleting `DesktopsGroup`'s closure, the only real one,
left the suite green. The floor counts call sites only.

The scan moves to `SettingsDisclosureAccessoryTests` with that
change: it watches `accessory:` closures, so it belongs beside
the ruling about that slot rather than beside what one file
draws — and the size suite had reached the 350-line ceiling.

Refs #1021
